[WIP] Add Date/Time dialog
Summary:
Import the date/time picker used in KTrip. It wraps the existing date/time picker components into a dialog. On Android it calls the native date/time picker instead.

The Android implementation is slightly messy since we apparently can't do JNI things directly from a QML plugin because it gets loaded in the wrong thread. As a workaround we have a linkable library on Android that forces JNI initialization in the right thread.

Finish

Test Plan: Ported KTrip to it

Reviewers: davidedmundson, vkrause, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29268