Add DateDialog component
This is the start for a date dialog that is able to make use of the native system dialog if available.

It currently only contains a fallback implementation based on the existing DatePopup component. An Android implementation will follow

API-wise it is modelled after the Qt.labs.platform dialogs, although only  a subset of that API is implemented