Add native date/time dialogs on Android
This is largely building on top of Nico's work from KTrip and
https://phabricator.kde.org/D29268. The native method registration
issue mentioned there doesn't seem to occur anymore though, this
works without any changes on the application side for Itinerary.
The additional indirection using a library is needed so androiddeployqt
does bundle the AAR file automatically.

On top of Nico's work this adds pre-selecting a date/time in the native
dialogs and integrates this with the DateInput and TimeInput controls.

# Conflicts:
#	src/dateandtime/CMakeLists.txt