Fix VideoMaximizeDelegate QT6
This fixes the VideoMaximizeDelegate download functionality for QT6.

The QT6 Video API removes a significant number of parameters and these have been worked around:
- The autoLoad parameter is gone and has been ported away from, to use the functionality the source should be set to QUrl() until the download is complete at which point the local url should be switched in.
- autoPlay has been re-implemented by calling play when there is a non-empty source

This should be transparent for anyone who wasn't overriding the download action and for anyone who was they will need to delete any useage of autoLoad.