Add support for initial values for unset date/times
When starting with an unset value this used the current date so far,
and you had to edit the date first before editing the time worked at all.

You can now customize the initial value (the default is still now), and
editing the time first on an unset value works as well.