AboutPage: Use '\n' because the format is forced to PlainText
To break the two paragraphs between KAboutData's otherText and the
copyright statement, a '<br/>' was used. At some point, the text format
was forced to PlainText so the tag was rendered in plaintext and didn't
put anything on a newline. Now the usual '\n' is used which works in
PlainText mode.