AlbumMaximizeComponent Media Manager
Setup play and pause so that a media manager can be used with AlbumMaximizeComponent.

This doesn't do anything if the play and pause actions are not overwritten but allows a consumer of the component to break in to perform supplemental actions before playing.

Done primarily to support network/neochat#641 and will be posting an MR for the NeoChat side.