Add ConfigurationsView
This is a rework of CategorizedSettings, visually it looks the same on
desktop and on mobile there is some small visual changes to match
plasma-settings.

The API is almost the same with the exeption that ConfigurationsView is
a plain QtObject while CategorizedSettings was a
Kirigami.PageRow. This means ConfigurationsView will take care of
opening itself with its open method while opening a CategorizedSettings
was the responsability of the application developer.

This is motivated by the recent breakage of CategorizedSettings on
desktop and the less recent breakage on mobile which were not fixable
with the same API. All the internals are not exposed so it is easy to in
the future change the visual if needed.