AboutPage: Remove windowing system
I find this to be a technical detail that's held over from days of old.
The windowing system (should) have little relevance for the application
and they're not usually built with advance knowledge of what windowing
system is in use.

I also don't think this is a particularly relevant detail for curious
users wishing to know how the application was made and who made it,
which is the purpose of the dialog.