Fix docs
The current docs layout with a dummy target for all things doesn't work. The QCH files can't be generated.

Instead the qhp definitions must be in the individual qdocconf files. The name of the qhp/qch files also must match the name of the qdocconf file (ECM assumes that), and because the qch files are installed into a global folder the names must be like "kirigamiaddonscomponents" instead of "components"