Introduce KirigamiApp
It's a library that provides a class to call in your main() function. It
helps clean up certain specificities that should generally be common to
all Kirigami apps and ensure we can keep our codebases clean while doing
everything we need to do for our apps to work as expected.