Rework KirigamiApp into a function
The initial design was changed so much during review, that I think it
didn't fit that well in the end.

Since it doesn't seem to have any users yet, I hope we can still change
it a bit.

This merge request moves all of the startup initialization code into a
single function, taking a user-constructed QGuiApplication, to enforce
it being called at the right time.

In my opinion, the difference between a KirigamiApp and a
KirigamiApp::App wasn't clear, and made the API confusing.
With this change, the initialization still looks very recognizable for
people comparing it to the Qt documentation, which I think is valuable.

Whether the application is running a QApplication or a QGuiApplication
is very important for people to know, as it changes the availability of
the qApp singleton. I think it is better to keep this part an explicit
decision.

I removed the load function for now, as it doesn't change that much over
the existing KLocalization API. If needed, we can add something like
it again later.