SearchPopupField: make sure that the search field is always centered in toolbars
The default Kirigami toolbars are now a little bit taller in order to be
consistent with all the other QWidget applications.
Tokodon had the search field not correctly centerd,
This because its toolbar has hardcoded curtom paddings.
This should be fixed, but in order to hav the current version keep
working, it needs also the searchpopupfield make less assumptions on
the toolbar height (right now assumes toolbar height == search field
height so it will always look centerd)

Regardless of the root item size, always center vertically the search
field and align the popup accordingly.