formcard/FormCardPage: automatically scroll to focused item
When using tab navigation to access elements of FormCardPage, the page
does not automatically scroll to make that element visible. This makes
keyboard-only navigation for sighted users very inconvenient.

ScrollablePage already has some machinery to scroll to particular
positions, so all we need to do here is install a handler for updates
to the activeFocusItem, and hook that into ScrollablePage's mechanism.

BUG: 466573